==> HealthInsuranceFraud/training_notebook/01_data_loading.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import sklearn

# Provider data

df_provider = pd.read_csv(r"..\data\Train-1542865627584.csv")

df_provider.head()

df_provider.info()

df_provider.describe()

df_provider.shape

df_provider.columns

df_provider['Provider'].nunique()

df_provider.isnull().sum()

# Beneficiary data

df_beneficiary = pd.read_csv(r"..\data\Train_Beneficiarydata-1542865627584.csv")

df_beneficiary.head()

df_beneficiary.shape

df_beneficiary.columns

df_beneficiary.isnull().sum()

df_beneficiary['DOD'].unique()

df_beneficiary['isDead'] = df_beneficiary['DOD'].notna().astype(int)


df_beneficiary.head()

df_beneficiary.drop(columns='DOD', inplace = True)

df_beneficiary.head()

df_beneficiary.isnull().sum()

# Inpatient data

df_inpatient = pd.read_csv(r"..\data\Train_Inpatientdata-1542865627584.csv")

df_inpatient.head()

df_inpatient.shape

df_inpatient.columns

df_inpatient.isnull().sum()

df_inpatient['has_attending_physician'] = df_inpatient['AttendingPhysician'].notna().astype(int)
df_inpatient['has_operating_physician'] = df_inpatient['OperatingPhysician'].notna().astype(int)
df_inpatient['has_other_physician'] = df_inpatient['OtherPhysician'].notna().astype(int)

df_inpatient['DeductibleAmtPaid'].head()

df_inpatient['DeductibleAmtPaid'].nunique()

df_inpatient['DeductibleAmtPaid'] = df_inpatient['DeductibleAmtPaid'].fillna(0)

df_inpatient.isnull().sum()

# making list of all the diagnosis codes columns
diagnosis_cols = [col for col in df_inpatient.columns if col.startswith("ClmDiagnosisCode")]
diagnosis_cols

df_inpatient[diagnosis_cols].isnull().sum()

df_inpatient['TotalDiagnosisCount'] = df_inpatient[diagnosis_cols].notnull().sum(axis=1)
df_inpatient['TotalDiagnosisCount']

row_index = 0  # For the first row (indexing is zero-based)
claim_details = df_inpatient.iloc[row_index]
print(claim_details)

procedure_cols = [col for col in df_inpatient.columns if col.startswith('ClmProcedureCode')]
total_procedure_codes = len(procedure_cols)
print(total_procedure_codes)

df_inpatient['TotalProcedureCount'] = df_inpatient[procedure_cols].notnull().sum(axis=1)
df_inpatient['TotalProcedureCount']

df_inpatient.isnull().sum()

df_inpatient['ProcedureDiagnosisRatio'] = df_inpatient['TotalProcedureCount'] / (df_inpatient['TotalDiagnosisCount']+1)
df_inpatient['ProcedureDiagnosisRatio']

df_inpatient['TotalProcedureCount'].describe()

df_inpatient['TotalDiagnosisCount'].describe()

df_inpatient['TotalProcedureCount'].quantile([0.95, 0.97, 0.99])  #95% of TotalProcedureCount is less than 2 and 5% is more than or equal to 2.

df_inpatient['TotalDiagnosisCount'].quantile([0.7,0.90,0.91,0.99])

max_diag = df_inpatient['TotalDiagnosisCount'].max()
df_inpatient['DiagnosisSlotsFull'] = (df_inpatient['TotalDiagnosisCount'] == max_diag).astype(int)
df_inpatient['DiagnosisSlotsFull'].sum()


df_inpatient['HighProcedure_95'] = (df_inpatient['TotalProcedureCount'] > 2).astype(int)


df_inpatient.columns

provider_features = df_inpatient.groupby('Provider').agg(
    TotalClaims=('Provider', 'count'),
    AvgDiagnosisCount=('TotalDiagnosisCount', 'mean'),
    AvgProcedureCount=('TotalProcedureCount', 'mean'),
    PctFullDiagnosis=('DiagnosisSlotsFull', 'mean'),
    PctHighProcedure=('HighProcedure_95', 'mean'),
    AvgProcDiagRatio=('ProcedureDiagnosisRatio', 'mean')
).reset_index()


provider_features.head()

inpatient_provider = provider_features.rename(columns={
    'TotalClaims': 'In_TotalClaims',
    'AvgDiagnosisCount': 'In_AvgDiagnosisCount',
    'AvgProcedureCount': 'In_AvgProcedureCount',
    'PctFullDiagnosis': 'In_PctFullDiagnosis',
    'PctHighProcedure': 'In_PctHighProcedure',
    'AvgProcDiagRatio': 'In_AvgProcDiagRatio'
})


# Outpatient data

df_outpatient = pd.read_csv(r"..\data\Train_Outpatientdata-1542865627584.csv")

df_outpatient.head()

df_outpatient.shape

df_outpatient.columns

df_outpatient.isnull().sum()

df_outpatient['AttendingPhysician'].nunique()

df_outpatient['has_attending_physician'] = df_outpatient['AttendingPhysician'].notna().astype(int)
df_outpatient['has_operating_physician'] = df_outpatient['OperatingPhysician'].notna().astype(int)
df_outpatient['has_other_physician'] = df_outpatient['OtherPhysician'].notna().astype(int)

df_outpatient['DeductibleAmtPaid'] = df_outpatient['DeductibleAmtPaid'].fillna(0)

diagnosis_cols = [col for col in df_outpatient.columns if col.startswith("ClmDiagnosisCode")]
diagnosis_cols

df_outpatient['TotalDiagnosisCount'] = df_inpatient[diagnosis_cols].notnull().sum(axis=1)
df_outpatient['TotalDiagnosisCount']

procedure_cols = [col for col in df_outpatient.columns if col.startswith('ClmProcedureCode')]
total_procedure_codes = len(procedure_cols)
print(total_procedure_codes)

df_outpatient['TotalProcedureCount'] = df_outpatient[procedure_cols].notnull().sum(axis=1)
df_outpatient['TotalProcedureCount']

df_outpatient['TotalProcedureCount'].unique()

df_outpatient['TotalDiagnosisCount'].quantile([0.7,0.90,0.91,0.99])

df_outpatient['TotalProcedureCount'].quantile([0.95, 0.97, 0.99])  #95% of TotalProcedureCount is less than 2 and 5% is more than or equal to 2.

df_outpatient['HasProcedure'] = (df_outpatient['TotalProcedureCount'] > 0).astype(int)


df_outpatient['Out_ProcedureDiagnosisRatio'] = df_outpatient['TotalProcedureCount'] / (df_outpatient['TotalDiagnosisCount']+1)
df_outpatient['Out_ProcedureDiagnosisRatio']

max_diag = df_outpatient['TotalDiagnosisCount'].max()
df_outpatient['DiagnosisSlotsFull'] = (df_outpatient['TotalDiagnosisCount'] == max_diag).astype(int)
df_outpatient['DiagnosisSlotsFull'].sum()


outpatient_provider = df_outpatient.groupby('Provider').agg(
    TotalClaims=('Provider', 'count'),
    AvgDiagnosisCount=('TotalDiagnosisCount', 'mean'),
    AvgProcedureCount=('TotalProcedureCount', 'mean'),
    PctFullDiagnosis=('DiagnosisSlotsFull', 'mean'),
    PctHasProcedure=('HasProcedure', 'mean'),
    AvgProcDiagRatio=('Out_ProcedureDiagnosisRatio', 'mean')
).reset_index()


outpatient_provider = outpatient_provider.rename(columns={
    'TotalClaims': 'Out_TotalClaims',
    'AvgDiagnosisCount': 'Out_AvgDiagnosisCount',
    'AvgProcedureCount': 'Out_AvgProcedureCount',
    'PctFullDiagnosis': 'Out_PctFullDiagnosis',
    'PctHasProcedure': 'Out_PctHasProcedure',
    'AvgProcDiagRatio': 'Out_AvgProcDiagRatio'
})


outpatient_provider.head()

outpatient_provider.describe()

provider_final = inpatient_provider.merge(
    outpatient_provider,
    on='Provider',
    how='outer'
)


provider_final = provider_final.fillna(0)


provider_final.shape


provider_final.head()

provider_final.describe()

provider_final = provider_final.merge(
    df_provider[['Provider', 'PotentialFraud']],
    on = 'Provider',
    how = 'left'
)

# covert yes to 1 and no to 0 in the provider table
provider_final['PotentialFraud'] = provider_final['PotentialFraud'].map({'Yes': 1, 'No': 0})

provider_final.to_csv("provider_fraud_features.csv", index=False)


df_provider['Provider'].nunique()

df_provider.columns

provider_final['PotentialFraud'].value_counts()


df_beneficiary.head()

df_beneficiary.columns
