==> HealthInsuranceFraud/training_notebook/05_final_training.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, confusion_matrix, ConfusionMatrixDisplay, average_precision_score
from sklearn.metrics import classification_report, roc_auc_score
from sklearn.pipeline import Pipeline
from sklearn.ensemble import RandomForestClassifier

df = pd.read_csv("../notebook/final_data.csv")
df.head()

df.columns

X = df.drop(columns=['Provider', 'PotentialFraud'])
y = df['PotentialFraud']

X.shape

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 42)

pipe_lr = Pipeline([
    ('scaler', StandardScaler()),
    ('lr', LogisticRegression(
        class_weight = 'balanced',
        max_iter = 1000,
        random_state = 42
    ))
])

pipe_lr.fit(X_train, y_train)

y_pred = pipe_lr.predict(X_test)

y_prob = pipe_lr.predict(X_test)
y_prob

y_prob = pipe_lr.predict_proba(X_test)[:, 1]
y_prob

report = classification_report(y_test, y_pred)
print(report)

"ROC-AUC:", roc_auc_score(y_test, y_prob)

average_precision_score(y_test, y_prob)

y.value_counts(normalize=True)

# Training model on RANDOM FOREST algorithm

pipe_rf = Pipeline([
    ('rf', RandomForestClassifier(
        n_estimators = 500,
        min_samples_leaf = 5,
        class_weight = 'balanced',
        random_state = 42,
        n_jobs = -1
    ))
])

pipe_rf.fit(X_train, y_train)

y_pred_rf = pipe_rf.predict(X_test)

y_prob_rf = pipe_rf.predict_proba(X_test)[:,1]
y_prob_rf

report_rf = classification_report(y_test, y_pred_rf)
print(report_rf)

"ROC-AUC:", roc_auc_score(y_test, y_prob_rf)

threshold = 0.4
y_pred_rf_tuned = (y_prob_rf >= threshold).astype(int)
print(classification_report(y_test, y_pred_rf_tuned))


import joblib

joblib.dump(pipe_rf, "fraud_model_rf.pkl")
